Refactor: Responses now contains ID to be referenced easier.
Still don't have a clue about a good way to manage IDs since it will be
mostly "private" usage. Since most portion just need a reference and not
the whole entity to make association i would need to have always an ID
in some way from the front-end, and to have it there i need to give it.
Well....